metric-generator/requirements.txt
-i https://pypi.org/simple
--extra-index-url https://artefact.skao.int/repository/pypi-all/simple
--extra-index-url https://nexus.engageska-portugal.pt/repository/pypi/simple
argparse==1.4.0
awkward1==1.0.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
awkward==1.4.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
docopt==0.6.2
future==0.18.2; python_version >= '2.6' and python_version not in '3.0, 3.1, 3.2, 3.3'
numpy==1.21.1
pyarrow==5.0.0
python-casacore==3.4.0
si-prefix==1.2.2
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
ska-sdp-dal-schemas==0.2.3
ska-sdp-dal==0.2.1
ska-ser-logging==0.4.1+0a677a67
spead2==3.3.2
tabulate==0.8.9
ska-sdp-cbf-emulator
requests
starlette
loguru
plotly